+++ packages/i18n/CHANGELOG.md
@@ -10,7 +10,13 @@ The following is a curated list of changes in the Enact i18n module, newest chan
 
 ### Changed
 
+- Vietnamese to be classified as a non-latin language
+
 ### Fixed
+
+## [1.8.0] - 2017-09-07
+
+Updated iLib to latest version
 
 ## [1.7.0] - 2017-08-23